Main/NATNetworkImpl.cpp: Signed/unsigned conversion issues. bugref:9790
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@85253 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-server/NATNetworkImpl.cpp
@@ -502,7 +502,7 @@ HRESULT NATNetwork::setLoopbackIp6(LONG aLoopbackIp6)
         if (static_cast<uint32_t>(aLoopbackIp6) == m->s.u32HostLoopback6Offset)
             return S_OK;
 
-        m->s.u32HostLoopback6Offset = (LONG)aLoopbackIp6;
+        m->s.u32HostLoopback6Offset = (uint32_t)aLoopbackIp6;
     }
 
     AutoWriteLock vboxLock(m->pVirtualBox COMMA_LOCKVAL_SRC_POS);
